pyproject.toml
[project]
name = "FastApiBasic"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "0xweirdlabs", email = "[redacted]@gmail.com"},
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.1",
    "prometheus-client>=0.20.0",
    "pydantic-settings>=2.3.3",

]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.2.2",
    "pylint>=3.2.3",
    "pytest-asyncio>=0.18.3",
    "httpx>=0.23.1",
]
[tool.pdm.scripts]
start-local = "python run_server.py local"
start-dev = "python run_server.py dev"
start-prod = "python run_server.py prod"
